# server/openapi.yaml
openapi: 3.1.0
info:
  title: EV2 Clandestinoz API
  version: 0.1.0
  description: |
    API del sistema de gestión de nightclub EV2 Clandestinoz.

    **Este archivo es el contrato entre backend, web y apps móviles.** Cualquier cambio
    de ruta o de forma de respuesta se refleja aquí en el mismo commit que lo introduce.

    ## Autenticación
    Salvo las rutas marcadas como públicas, toda petición requiere
    `Authorization: Bearer <access_token>`. El token de acceso dura 15 minutos; se renueva
    con `POST /api/auth/refresh` usando el refresh token, que es de **un solo uso**
    (al canjearlo se revoca y se emite uno nuevo).

    ## Errores
    Todos los errores tienen la misma forma (`ErrorResponse`). Los códigos usados son
    400 validación, 401 sin autenticar, 403 sin permiso, 404 no existe, 409 conflicto de
    estado, 422 regla de negocio, 429 límite de peticiones, 500 interno, 501 no implementado.

    ## Dinero
    Los importes son cadenas decimales con dos decimales (`"1500.00"`) para no perder
    precisión, y siempre van acompañados de `currency` (`MXN` o `USD`). Los precios se
    calculan **siempre en el servidor**.

    ## Precios de reservación
    Una mesa se reserva por **toda la noche**, que corresponde a un evento del calendario.
    El precio es `base de la zona + (personas extra x precio del boleto de esa noche) +
    extras`. Cada zona incluye un número de boletos y admite un tope de extras (por
...

...

    ## Flirt
...

    ## Empleados y dinero
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/nightclubs/{nightclubId}:
    get:
      operationId: getNightclubsByNightclubId
      tags: [Club]
      summary: Perfil del club
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      responses:
        '200':
          description: Perfil
          content:
            application/json:
              schema:
                type: object
                properties:
                  nightclub: { $ref: '#/components/schemas/Nightclub' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/nightclubs/{nightclubId}/payment-providers:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
    get:
      operationId: getPaymentProviders
      tags: [Pagos manuales]
      summary: Con qué se puede cobrar hoy
      description: |
        El gerente ve qué pasarelas están configuradas y **qué llaves faltan**, para que
        conectar Stripe o Mercado Pago sea una lista de pendientes y no una adivinanza
        (`docs/PAGOS_SETUP.md`). El cliente ve solo las que ya funcionan, con la llave
        pública que su navegador necesita.

        Las llaves **secretas viven en variables de entorno, nunca en la base de datos**, y
        **nunca salen por esta ruta**. Mientras no haya ninguna pasarela, el club cobra con
        pagos manuales, que no dependen de ninguna cuenta externa.
      responses:
        '200':
          description: Estado de las pasarelas
          content:
            application/json:
              schema:
                type: object
                properties:
                  manual_available: { type: boolean }
                  any_provider_configured:
                    type: boolean
                    description: Solo para el gerente.
                  providers:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentProviderStatus' }
  /api/nightclubs/{nightclubId}/manual-payment-options:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
    get:
      operationId: listManualPaymentOptions
      tags: [Pagos manuales]
      summary: A dónde puede pagar el cliente
      parameters:
        - name: include_inactive
          in: query
          description: Solo el gerente ve las retiradas.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Formas de pago
          content:
            application/json:
              schema:
                type: object
                properties:
                  options:
                    type: array
                    items: { $ref: '#/components/schemas/ManualPaymentOption' }
    post:
      operationId: createManualPaymentOption
      tags: [Pagos manuales]
      summary: Cargar una forma de pago (gerente)
      description: |
        `destination` es la cuenta, CLABE o usuario al que el cliente envía el dinero, y se
        guarda en claro a propósito: a diferencia de la cuenta de un empleado (cifrada en
        la migración 007), esta es la cuenta receptora del club y su función es mostrarse a
        quien va a pagar. Una transferencia sin destino se rechaza; el efectivo no lo pide.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualPaymentOptionCreate' }
      responses:
        '201':
          description: Forma de pago creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  option: { $ref: '#/components/schemas/ManualPaymentOption' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/nightclubs/{nightclubId}/manual-payment-options/{optionId}:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
      - $ref: '#/components/parameters/OptionId'
    put:
      operationId: updateManualPaymentOption
      tags: [Pagos manuales]
      summary: Editar una forma de pago (gerente)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualPaymentOptionUpdate' }
      responses:
        '200':
          description: Forma de pago actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  option: { $ref: '#/components/schemas/ManualPaymentOption' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: retireManualPaymentOption
      tags: [Pagos manuales]
      summary: Retirar una forma de pago (gerente)
      description: 'Se marca inactiva, nunca se borra: los pagos ya declarados la referencian.'
      responses:
        '200':
          description: Retirada
          content:
            application/json:
              schema:
                type: object
                properties:
                  option: { $ref: '#/components/schemas/ManualPaymentOption' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/nightclubs/{nightclubId}/manual-payments:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
    get:
      operationId: listManualPayments
      tags: [Pagos manuales]
      summary: Cola de pagos por revisar (gerente)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [declared, confirmed, rejected, cancelled] }
        - name: method
          in: query
          schema: { type: string, enum: [cash, card_terminal, zelle, cash_app, bank_transfer, spei] }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Pagos
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: '#/components/schemas/ManualPayment' }
                  awaiting_review:
                    type: array
                    items:
                      type: object
                      properties:
                        currency: { $ref: '#/components/schemas/Currency' }
                        count: { type: integer }
                        total: { $ref: '#/components/schemas/Money' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: declareManualPayment
      tags: [Pagos manuales]
      summary: Declarar que ya se pagó
      description: |
        **No cobra nada.** Deja el asiento del libro en `pending_manual` a la espera de que
        el gerente lo confirme contra el estado de cuenta. Una transferencia exige folio:
        sin él no hay forma de comprobarla. Un cliente solo declara sus propios cobros; el
        personal puede declarar por otra persona con `on_behalf_of`. Idempotente por
        `client_request_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualPaymentDeclare' }
      responses:
        '201':
          description: Pago declarado
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment: { $ref: '#/components/schemas/ManualPayment' }
        '200':
          description: Ya estaba declarado (idempotencia)
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment: { $ref: '#/components/schemas/ManualPayment' }
                  idempotent: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/nightclubs/{nightclubId}/manual-payments/mine:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
    get:
      operationId: listMyManualPayments
      tags: [Pagos manuales]
      summary: Mis pagos declarados
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Pagos del cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: '#/components/schemas/ManualPayment' }
...
  /api/nightclubs/{nightclubId}/manual-payments/{paymentId}:
    parameters:
      - $ref: '#/components/parameters/NightclubId'
      - $ref: '#/components/parameters/PaymentId'
    get:
      operationId: getManualPayment
      tags: [Pagos manuales]
      summary: Ver un pago
      description: |
        El cliente ve su declaración y en qué va; el gerente ve además quién la hizo, quién
        la revisó y el asiento contra el que se aplica. Un pago ajeno responde 404.
      responses:
        '200':
          description: Pago
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment: { $ref: '#/components/schemas/ManualPayment' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso de 15 minutos obtenido en login o refresh.
    agentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: |
        Llave propia del agente local del POS. Autentica una maquina, no una persona:
        no hereda permisos de ningun usuario y solo sirve en las rutas /api/pos/agent/*.
...
  responses:
...
    Unauthorized:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...